refactor mockup data into folder
