+++ prisma/data-mockup/eventAddress.js
@@ -1,0 +1,21 @@
+modules.exports = eventAddress = [
+    { eventId: 1, address: "123 Main Street, Bangkok", provinceId: 1, lat: 13.7563, long: 100.5018 },
+    { eventId: 2, address: "456 Elm Street, Phuket", provinceId: 2, lat: 7.8804, long: 98.3923 },
+    { eventId: 3, address: "789 Oak Street, Chiang Mai", provinceId: 3, lat: 18.7903, long: 98.9847 },
+    { eventId: 4, address: "101 Maple Street, Ayutthaya", provinceId: 4, lat: 14.3588, long: 100.5582 },
+    { eventId: 5, address: "202 Pine Street, Samut Prakan", provinceId: 5, lat: 13.6529, long: 100.4887 },
+    { eventId: 6, address: "303 Birch Street, Nakhon Pathom", provinceId: 6, lat: 13.8196, long: 100.0443 },
+    { eventId: 7, address: "404 Cedar Street, Songkhla", provinceId: 7, lat: 7.1898, long: 100.5951 },
+    { eventId: 8, address: "505 Fir Street, Udon Thani", provinceId: 8, lat: 17.415, long: 102.7855 },
+    { eventId: 9, address: "606 Spruce Street, Khon Kaen", provinceId: 9, lat: 16.4467, long: 102.833 },
+    { eventId: 10, address: "707 Redwood Street, Nakhon Ratchasima", provinceId: 10, lat: 14.9799, long: 102.0977 },
+    { eventId: 11, address: "808 Walnut Street, Surat Thani", provinceId: 11, lat: 9.1347, long: 99.3331 },
+    { eventId: 12, address: "909 Pine Street, Chiang Rai", provinceId: 12, lat: 19.9072, long: 99.8308 },
+    { eventId: 13, address: "111 Maple Street, Phra Nakhon Si Ayutthaya", provinceId: 13, lat: 14.3559, long: 100.5679 },
+    { eventId: 14, address: "121 Elm Street, Nakhon Sawan", provinceId: 14, lat: 15.672, long: 100.1313 },
+    { eventId: 15, address: "131 Oak Street, Rayong", provinceId: 15, lat: 12.6819, long: 101.2813 },
+    { eventId: 16, address: "141 Cedar Street, Nakhon Nayok", provinceId: 16, lat: 14.2037, long: 101.214 },
+    { eventId: 17, address: "151 Fir Street, Trang", provinceId: 17, lat: 7.5596, long: 99.6107 },
+    { eventId: 18, address: "161 Birch Street, Sakon Nakhon", provinceId: 18, lat: 17.1628, long: 104.1479 },
+    { eventId: 19, address: "171 Redwood Street, Pattaya", provinceId: 19, lat: 12.9316, long: 100.8828 },
+]
