+++ prisma/data-mockup/eventImage.js
@@ -1,0 +1,54 @@
+modules.exports = eventImage = [
+    {
+        eventId: 1,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 1,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 1,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 1,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 1,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 2,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 2,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 2,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 2,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 2,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 2,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 3,
+        image: "https://picsum.photos/600/600",
+    },
+    {
+        eventId: 4,
+        image: "https://picsum.photos/600/600",
+    },
+]
