+++ prisma/data-mockup/eventType.js
@@ -1,0 +1,19 @@
+module.exports = eventType = [
+    {
+        name: "Cultural ",
+    },
+    {
+        name: "Community ",
+    },
+    {
+        name: "Sports ",
+    },
+    {
+        name: "Music ",
+    },
+    {
+        name: "Art ",
+    },
+    {
+        name: "Historical  ",
+    },]
