+++ prisma/data-mockup/events.js
@@ -1,0 +1,249 @@
+modules.exports = events = [
+    {
+        organizerInformationId: 1,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Watarun",
+        description: "the best wat in thailand",
+        startDate: new Date("2024-10-03"),
+        endDate: new Date("2024-12-03"),
+        eventTypeId: 1,
+        timePeriod: "08:00 AM - 05:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 2,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Adventure Fest",
+        description: "Explore the wonders of nature",
+        startDate: new Date("2024-08-10"),
+        endDate: new Date("2024-12-10"),
+        eventTypeId: 2,
+        timePeriod: "09:00 AM - 06:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 3,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Culinary Delight",
+        description: "A feast for food lovers",
+        startDate: new Date("2024-07-12"),
+        endDate: new Date("2024-11-12"),
+        eventTypeId: 3,
+        timePeriod: "11:00 AM - 08:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 4,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Tech Expo",
+        description: "Innovations that shape the future",
+        startDate: new Date("2024-09-08"),
+        endDate: new Date("2024-05-08"),
+        eventTypeId: 4,
+        timePeriod: "10:00 AM - 07:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 5,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Fusion Beats",
+        description: "A musical journey across genres",
+        startDate: new Date("2024-11-15"),
+        endDate: new Date("2024-12-15"),
+        eventTypeId: 5,
+        timePeriod: "03:00 PM - 11:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 6,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Artistic Vibes",
+        description: "Celebrating creativity in all forms",
+        startDate: new Date("2024-06-09"),
+        endDate: new Date("2024-10-09"),
+        eventTypeId: 6,
+        timePeriod: "09:30 AM - 06:30 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 1,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Mindful Retreat",
+        description: "Embrace tranquility and peace",
+        startDate: new Date("2024-04-11"),
+        endDate: new Date("2024-06-11"),
+        eventTypeId: 1,
+        timePeriod: "10:00 AM - 06:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 2,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Gastronomy Gala",
+        description: "A culinary journey around the world",
+        startDate: new Date("2024-12-14"),
+        endDate: new Date("2024-12-30"),
+        eventTypeId: 2,
+        timePeriod: "12:00 PM - 09:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 3,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Code Confluence",
+        description: "Where coding brilliance meets",
+        startDate: new Date("2024-03-13"),
+        endDate: new Date("2024-05-13"),
+        eventTypeId: 2,
+        timePeriod: "01:30 PM - 09:30 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 4,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Fashion Fiesta",
+        description: "A showcase of style and trends",
+        startDate: new Date("2024-02-16"),
+        endDate: new Date("2024-03-16"),
+        eventTypeId: 3,
+        timePeriod: "04:00 PM - 11:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 5,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Health Harmony",
+        description: "Wellness for mind and body",
+        startDate: new Date("2024-05-10"),
+        endDate: new Date("2024-10-10"),
+        eventTypeId: 4,
+        timePeriod: "09:00 AM - 05:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 6,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Cultural Carnival",
+        description: "Diverse traditions under one roof",
+        startDate: new Date("2024-07-12"),
+        endDate: new Date("2024-09-12"),
+        eventTypeId: 5,
+        timePeriod: "11:30 AM - 08:30 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 1,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Tech Summit",
+        description: "Exploring the future of technology",
+        startDate: new Date("2024-09-08"),
+        endDate: new Date("2024-12-08"),
+        eventTypeId: 6,
+        timePeriod: "10:00 AM - 06:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 1,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Watarun",
+        description: "the best wat in thailand",
+        startDate: new Date("2024-10-03"),
+        endDate: new Date("2024-11-03"),
+        eventTypeId: 1,
+        timePeriod: "08:00 AM - 05:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 2,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Adventure Fest",
+        description: "Explore the wonders of nature",
+        startDate: new Date("2024-08-10"),
+        endDate: new Date("2024-09-10"),
+        eventTypeId: 2,
+        timePeriod: "09:00 AM - 06:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 3,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Culinary Delight",
+        description: "A feast for food lovers",
+        startDate: new Date("2024-07-12"),
+        endDate: new Date("2024-12-12"),
+        eventTypeId: 3,
+        timePeriod: "11:00 AM - 08:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 4,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Tech Expo",
+        description: "Innovations that shape the future",
+        startDate: new Date("2024-09-08"),
+        endDate: new Date("2024-10-08"),
+        eventTypeId: 4,
+        timePeriod: "10:00 AM - 07:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 5,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Fusion Beats",
+        description: "A musical journey across genres",
+        startDate: new Date("2024-11-15"),
+        endDate: new Date("2024-12-15"),
+        eventTypeId: 5,
+        timePeriod: "03:00 PM - 11:00 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+    {
+        organizerInformationId: 6,
+        coverImage: "https://picsum.photos/600/300",
+        title: "Artistic Vibes",
+        description: "Celebrating creativity in all forms",
+        startDate: new Date("2024-06-09"),
+        endDate: new Date("2024-10-09"),
+        eventTypeId: 6,
+        timePeriod: "09:30 AM - 06:30 PM",
+        isYearly: false,
+        isPublish: true,
+        isExpire: false,
+    },
+]
