+++ prisma/data-mockup/index.js
@@ -1,0 +1,11 @@
+const seedData = {
+    province: require("./provinces"),
+    eventType: require('./eventType'),
+    user: require('./users'),
+    organizerInformation: require('./organizerInformations'),
+    event: require('./events'),
+    eventAddress: require('./eventAddress'),
+    eventImage: require('./eventImage')
+}
+
+module.exports = seedData
