+++ prisma/data-mockup/organizerInformations.js
@@ -1,0 +1,54 @@
+const CORPORATION = {
+    INDIVIDUAL: "INDIVIDUAL",
+    CORPORATION: "CORPORATION",
+}
+
+const organizeImage = "https://download-th.com/wp-content/uploads/2020/12/TAT.jpg"
+const identityImage = "https://www.dmv.ca.gov/portal/uploads/2020/06/realid_img-1024x656.jpg"
+
+
+
+modules.exports = organizerInformation = [
+    {
+        userId: 1,
+        officialName: "Top Thailand",
+        corporation: CORPORATION.CORPORATION,
+        identityCopyImage: identityImage,
+        profileImage: organizeImage,
+    },
+    {
+        userId: 2,
+        officialName: "Pat Thailand",
+        corporation: CORPORATION.CORPORATION,
+        identityCopyImage: identityImage,
+        profileImage: organizeImage,
+    },
+    {
+        userId: 3,
+        officialName: "japan Thailand",
+        corporation: CORPORATION.CORPORATION,
+        identityCopyImage: identityImage,
+        profileImage: organizeImage,
+    },
+    {
+        userId: 4,
+        officialName: "guide Thailand",
+        corporation: CORPORATION.CORPORATION,
+        identityCopyImage: identityImage,
+        profileImage: organizeImage,
+    },
+    {
+        userId: 5,
+        officialName: "bank Thailand",
+        corporation: CORPORATION.CORPORATION,
+        identityCopyImage: identityImage,
+        profileImage: organizeImage,
+    },
+    {
+        userId: 6,
+        officialName: "pete Thailand",
+        corporation: CORPORATION.CORPORATION,
+        identityCopyImage: identityImage,
+        profileImage: organizeImage,
+    },
+]
